Change empty message to be more directive
Currently the tool just says "Empty MM context is loaded" on start,
which is *true* but doesn't a hint to a new user of what to do next.

Since this is one of the *first* things a new user will see, I
think it's important to give them more specific directions.
In this case, tell them specifically to select a database to load
The tool isn't very useful without a database, so I think we can be
prescriptive at this point. Anyone who can use this tool with an
empty database is advanced enough to not need prompting :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>