Modify 2+2=4 proof to "clean up as you go"
Modify the proof of 2+2=4 to "clean up as you go",
in this case, to show that 4 is equal to the expression
we generated from 3+1. Cleaning up as you go is probably
a good practice anyway. It also makes explaining the proof,
and applying metamath-lamp, a little easier.

This also shows an example of duplicating a statement.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>