Change "statement" to "step" in UI
We recently agreed that it'd be clearer in the UI if we said
that there were many *steps*, and each step contained a *statement*.
This commit modifies the UI to more consistently say "step" when
step is what was intended. Renaming "statement" to "step" everywhere
in the UI would be incorrect, as in many cases "statement" is intended.
I also changed "id" to "label" in a few places, again, for consistency.

There are probably a few missed spots, which will have to be fixed
by future commits, but this at least gets things started.

This does *not* try to rename anything in the code, just the UI.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>