Shorten context summary at top for smartphones
The context summary at the top is too long for smartphones
even when you're stopped just before theorems with short names
like '2p2e4'.

This commit shortens the displayed text so that it provides
the same information but with less text. It removes the ";"
after "Loaded", and it removes the unnecessary word
"stopped" (this was already shown when you selected it, and
"before" and "after" already convey the same information).

The result is that on '2p2e4' and 'reccot' it only requires
*one* line instead of two, so the user doesn't need to configure
anything for it to take up less space. Smartphone displays
are small, making display space a premium, and we especially
don't want to distract people who are just starting to use the tool.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>