Rename icon "substitution" to "replacement" in UI
The word "substitution" is confusing, because it has multiple meanings.
Is it the fundamental Metamath operation? The "A with arrow" icon?

This commit changes the UI so that the "A with arrow" icon
is *always* described as "replacement", never as "substitution".
This action already used the term "replacement" in some places
(e.g., "Replace with?"), but not in others.
Thus, it made sense to switch this icon so it consistently
uses the term replacement.

This commit does *not* try to rename anything in the code itself.
The user doesn't care what the code calls it. The code could be
changed later to match this UI naming convention, but there's no
urgent need to do that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>