Clarify step type dropdown
Up to this point the step type dropdown has only shown a
single letter (P, H, and recently G).

This commit modifies the dropdown to give a slightly longer
explanation of what each letter means
(if the user doesn't know or remember).
It also records the current state, so the current state is
highlighted in the dropdown.
Hopefully this makes the UI just a little more discoverable
and clearer to new users.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>